Issue 94: Move dummy response message helpers to separate class.
